requirements.txt
kafka-python-ng==2.2.3
mysql-connector-python==9.1.0
requests==2.32.3
six==1.17.0